requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-filter==21.1
django-js-asset==1.2.2
django-phone-field==1.8.1
django-phone-verify==0.1.1
django-phonenumber-field==5.2.0
django-recaptcha==2.0.6
djangorestframework==3.12.4
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
httplib2==0.20.1
idna==3.3
Markdown==3.3.4
oauthlib==3.1.1
phonenumbers==8.12.36
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.2.0
pyparsing==2.4.7
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
twilio==7.3.0
urllib3==1.26.7
